Revert "Add a new predicate to skip messages that are equal to the previous one"
With the upcoming name change of `OnlyIfPrevious` to `WithPrevious`, it
doesn't look like `filter(WithPrevious(lambda prev, new:
prev != new))` is so ugly and verbose to justify having `ChangedOnly` as
a special case. Specially since the name is a bit confusing, and not
clear enough about what it does. The verbose alternative make it more
clear that it's comparing to the previous message.

This reverts commit 2b9541ceba62e5e5928ad0905f61c3495bc15280.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>